Moved csv function to result_api.py
