+++ rgpvApi/__init__.py
@@ -1,3 +1,3 @@
 from .info_api import info
 from .result_api import result
-from .csv_result_api import csvresults
+from .result_api import csvresults
